first commit working on owl generation fixed problem with database split test cases vinay & heinz three graphs languag, variables, equations A first GUI for choosing qudt terms and thus their IRIs. fix when buttons become visible
IRI can now be defined for each variable

Tried a visualisation, but does not work. Too big.

indices and equations added

new simplified version

2023-03-23 played with how to put the expressions into the ontology. The first version was as a list using the ontologised ProMo language. Turned out to be unnecessarily complicated. Adding it as a Literar is essentially the same. The code has still both versions controlled by one global variable <render_expression_to_list>
There is also a SPARQL example <test_sparql.py> in the test/tests_heinz branch

An issue with the numbering of the equations has been detected in the <ontology_container.py> and marked as TODO

new task for generating the RDF graph and the ttl file was moved into the ontology directory.

Ontology writing and reading established
missing -- compiler and code editor

changes in the RDF graph